combine misc and hepl cogs
